Move the test:integration tests aside
These are now moved to a separate test so that WebMock has not been
loaded and would therefore interfere with actually hitting the
hosted verifier. By having them in separate tasks means that they
are run separately and therefore WebMock is not loaded when doing
the integration tests. Run them all with:

$ rake test test:integration

or

$ rake test
$ rake test:integration